requirements.txt
asgiref==3.2.3
Django==3.0.3
django-cors-headers==3.2.1
django-filter==2.4.0
djangorestframework==3.11.0
gunicorn==20.0.4
pycodestyle==2.6.0
pytz==2019.3
sqlparse==0.3.0
toml==0.10.2
whitenoise==5.2.0
